network/setup: Do not override existing data in the pod iface cache
During the discovery stage, the pod-interface-cache is used to store pod
related data.

When the data is stored in the cache, it overrides anything that may
have existed there beyond the actual data which is stored.

In order to use the same back-end persisted storage with the
pod-iface-state, this change assured that existing data is not
overridden.

Signed-off-by: Edward Haas <edwardh@redhat.com>