Add ksmConfiguration to enable ksm in the nodes
With ksmConfiguration cluster admins are able to define a set of nodes
in which ksm will be enabled.
The nodes will be filtered by label with the LabelSelector.
If a node matches the labels and has the ksm available, the ksm in that node
will be enabled. Alongside the enabling of the ksm, an annotation will be added
to the node.
This annotation marks nodes where the ksm has been enabled by the virt-handler,
allowing kv to disable it in case a subsequent change in the selector
will cause a node to no longer be included.

Signed-off-by: fossedihelm <ffossemo@redhat.com>