virt-controller: Refactor syntactically the multus CNI creation
Instead of passing the VMI object to the multus CNI generation
functions, pass only the needed information: namespace, interfaces and
networks.

With the hotplug supporting function taking the same approach, it will
be possible to pass in to the multus annotation generation functions
filtered interfaces/networks, avoiding the need to embed into those
functions logic which do not belong there.

Signed-off-by: Edward Haas <edwardh@redhat.com>