virt-controller, vmi: Update multus annotations on hot-unplug
The VMI controller is detecting if interfaces need to be hotplug based
on the spec networks and current interfaces status.
With the introduction of the `absent` state, interfaces marked for
removal should not be considered for hotplug.

Adjusted the current handling of the hotplug to ignore
interfaces/networks that are marked for removal.

In order to also support hot-unplug, a condition for which the multus
annotation is re-generated is added. When there is at least one
interface marked for removal, the annotations are regenerated and
updated.

In addition, old VMIs which have their pod interfaces names ordered (e.g. net1) are not removed from the multus annotation.

Signed-off-by: Edward Haas <edwardh@redhat.com>