virt-launcher: Detach interfaces marked for removal
Network interfaces that appear on the VMI spec with state set as
`absent` are targeted for removal.

Detect interfaces that exist in the domain and are marked as `absent`.
Detach them from the domain.

Do not detach interfaces which have target device different from the
expected hashed name (tap<hash based on network-name>).
Interfaces marked as `absent` are now ignored when calculating which
interfaces to hotplug.

Signed-off-by: Edward Haas <edwardh@redhat.com>