Avoid test failure that occurs when lost+found exists
lost+found is a special directory in the root directory of a mounted
file system  to help recover files lost to file system corruption and
is restricted to root.

When it exists, our exported tarballs will contain it, and when we try
to unpack it in the download pod where lost+found exists too, we will
have a permissions error for trying to modify the permissions and
contents of this directory with the unprivileged download pod.

Avoid this problem by only extracting the file that is going to be
checksummed for comparison, the other files are not checked.

Signed-off-by: Maya Rashish <mrashish@redhat.com>