network, setup: Relocate the pod-iface discovery
The pod interface name information has been collected and updated in the
nic structure at the discovery step.

The information in question is needed for other tasks (e.g. should the
setup be executed on the specific iface or not).

Therefore, place another hookpoint (function pointer) at the ConfigState
run method to implement an early update (or filtering) on the nic list.

Signed-off-by: Edward Haas <edwardh@redhat.com>