handle nonRoot -> root FG transition (#2353)
Up to version v0.58, Kubevirt was consuming
a featuregate named nonRoot, with
v0.59 it got renamed as Root with the opposite
logic and so we implemented
https://github.com/kubevirt/hyperconverged-cluster-operator/pull/2273
as an internal workaround for the time being.

Let's now align with Kubevirt featureGate name
for the sake of consistency with a proper deprecation
plan.

`nonRoot` is still exposed on HCO but it's
deprecated and a new `Root` FG is introduced.

The mutating webhook will take care
of the conversion.

A Common Expression Language rule on the CRD
will enforce the consistency of the two antithetic
values.

In the next release `nonRoot` will be removed.

Adding unit and e2e tests.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=2174859

Signed-off-by: stirabos <stirabos@redhat.com>