Fix local execution on OCP (#2355)
When we deploy locally, HCO binary
is executed without a pod (by design) so we cannot
get a reference to its deployment and to
the owning CSV starting from the pod we are
running in.
On the other side, when deploying locally,
we know in advance the name of the deployment
object (with replica=0) and the expected
namespace from an environmental variable.
Let's get the deployment object from there.

Signed-off-by: stirabos <stirabos@redhat.com>