build virtctl also for arm64
When we build on linux on amd64 we are already
building virtctl also for darwin and windows
OSs but still only for amd64.
But client architecture could differ the cluster ones.

kubectl for instance is shipped for linux, darwin
and windows in both amd64 and arm64 architectures.
Check as a reference:
- https://dl.k8s.io/release/v1.27.2/bin/linux/amd64/kubectl
- https://dl.k8s.io/release/v1.27.0/bin/linux/arm64/kubectl
- https://dl.k8s.io/release/v1.27.2/bin/darwin/amd64/kubectl
- https://dl.k8s.io/release/v1.27.2/bin/darwin/arm64/kubectl
- https://dl.k8s.io/release/v1.27.2/bin/windows/amd64/kubectl.exe
- https://dl.k8s.io/release/v1.27.2/bin/windows/arm64/kubectl.exe

Let's start taking the same approach for virtctl.

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>