api: Drop the `kubevirt.io/interface` resource name API
The network interfaces reservation API is dropped in favor of a simple
default.
Basic tests show that there is no strong justification to expose the
knob when the memory and boot time effect is negligible.

The default reservation value for interfaces is set to 4.

If users explicitly define a domain to have no interface at all,
no reservation is set.
Therefore, reservation is set only if at least one interface is defined
in the domain and the number of defined interfaces is below 4.

This reverts commit 1fcf489a27eba10de17fae7696209833cef18a20.

Signed-off-by: Edward Haas <edwardh@redhat.com>