virt-launcher, nichotplug: Use only the auto-generated domain devices
When reserving network interfaces, the domain configuration is created
in two steps:
- Create the domain configuration with placeholder interfaces.
- Create the domain configuration based on the auto-generated
  configuration from the previous step but without the placeholder
  interfaces.

When retrieving the auto-generated configuration, it is missing the
`qemu:commandline` section. This in practice drops configuration that
supports SLIRP and potentially other features.

Therefore, do not use the whole generated domain configuration for the
second step, just use the devices section.

Signed-off-by: Edward Haas <edwardh@redhat.com>