Make sure that the number of VMs to evacuate is never negative
Lately we encountered a bug where the number of VMs to evacuate
was negative. This was probably caused by some kind of a race.

For example, if there are X migrations running on a node,
and ParallelOutboundMigrationsPerNode is then changed to 1,
than the free spots of migrations to create in this node is:
ParallelOutboundMigrationsPerNode-X = 1 - X < 0.

This should never happen. Therefore, this variable is always
set to 0 if negative.

Signed-off-by: Itamar Holder <iholder@redhat.com>