virt-launcher, hotplug: No reservation with placePCIDevicesOnRootComplex
When using the "kubevirt.io/placePCIDevicesOnRootComplex" VMI
annotation, network interface reservation is not used.

Placing PCI devices on the root complex moves the allocation of PCI
controllers to be owned & managed by kubevirt. This collides with the
reservation implementation logic which offloads the ownership and logic
to libvirt.

If specific requirements will arise in the future, this exception will
be reconsidered.

Signed-off-by: Edward Haas <edwardh@redhat.com>