Set the deletion grace period to 1 on forced VM restarts
In order to allow a swift and safe restart, set the deletion grace
period to 1 on the pod delete when a force restart is requested.

One is the shortest safe restart period where we don't risk a
split-brain scenario for VMs, since Zero effectively deletes the pod
right away.

This leads to the following outcome:
 * If a guest OS is stuck, people get a swift restart.
 * Users don't risk data corruption.
 * Reserving a force pod delete for situations where the cluster is
   unhealthy.

Signed-off-by: Roman Mohr <rmohr@google.com>