virt-launcher, notify: Fix seg-fault due to too early metadata handling
It is expected to first receive notifications/events from libvirt and
create the initial domain object (which is used throughout the
go-routine).
But it seems there are incidents where the metadata cache notification
arrives before and crashes the virt-launcher.

Fixed by ignoring such early notifications on the metadata cache.
Only after the first libvirt event arrives the processing of metadata
events is handled.

Fixes: https://github.com/kubevirt/kubevirt/issues/9839

Signed-off-by: Edward Haas <edwardh@redhat.com>