migration-controller: label VMI with CPU limits
In CPU hotplug scenario virt-handler needs to know which
VMI pod is the right one for CPU hotplug action. There can be a potential
race condition when user triggers live migration and does
CPU hotplug at the same time. In such case, if dedicated CPU was configured
the target pod can be created with insufficient resources (these before the hotplug changes)

The solve it, virt-handler will compare the CPU topology of the VMI
with the labeled resource limits. If it exceeds the limits virt-handler
won't do the hotplug.

Passing this infornation via label is needed since
virt-handler doesn't have access to pod object

This label will take place as part of live-migration flow
in migration contoller, at the phase where the target pod
is determined as ready.

The case with shared CPU considered less relevant ATM because virt-handler
can't know whether the CPU overcommit ratio changed by comparing the CPU request
time with the CPU topology.

Signed-off-by: enp0s3 <ibezukh@redhat.com>
Co-authored-by: Vladik Romanovsky <vromanso@redhat.com>