Handle cluster level eviction strategy on SNO as a special case (#2376)
With #2071 we introduced a knob to let the cluster
admin configure a default EvictionStrategy at
cluster level.
The default value there was `LiveMigrate` for all
the kinds of clusters but this could potentially
introduce side effects (alerts and potentially
also inability to upgrade the cluster without
manually recovery actions) on SNO clusters.

Let's set a different default value (`None`) for
cluster level eviction strategy when we detect
that we are on a cluster with a single worker node
where `LiveMigrate` is not an option.

The code will still honour explicit user choice if there.

Add unit and e2e tests.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=2212289

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>