virt-handler, network setup: Do not create podNIC for SR-IOV networks
Kubevirt expects SR-IOV interfaces to be set with userspace
driver (vfio-pci) in order to attach them to VMs.
The interfaces are passthroughed to the VM using Libvirt API.
Thus SR-IOV networks should not be part of the network setup.

Also, fix netconf unit test that fails because there was no
matching interface in the VMI spec.

Signed-off-by: Or Mergi <ormergi@redhat.com>