hotunplug: combine hotunplug flow with the plug
This commit fixes several review comments.
1. combine hotunplug flow with the plug
2. remove form ConfigState the podIfaceName cache. On each hotunplug,
configState will enter the virt-launcher net namespace.
The expesive case is when unplugging an ordinal interface, then on each
reconcile of the virt-handler the virt-launcher's net ns will be entered.
But it is an edge case that doesn't worth such a complex solution.
In a follow-up we should block/remove Absent state from such interfaces.

Signed-off-by: Alona Paz <alkaplan@redhat.com>