Network, setup: Remove redundant lookup of interface by network
Following PRs #9268 and #9890, we now look up an interface by its network name
in `getPhase1NICs()` and in `getPhase2NICs()`.

The same lookup operation also occurs at `newPodNIC()` which is called from both
`getPhase1NICs()` and `getPhase2NICs()`.

Pass the interface found in `getPhase1NICs()` and `getPhase2NICs()`
down the stack to `newPodNIC()`, so it could be used instead of performing the
same operation again.

Signed-off-by: Orel Misan <omisan@redhat.com>