Use formatted archname in push-container-manifest
`bazel-push-images.sh` puts the digests files in paths based on the
formatted arch names.
The same does not happen in `push-container-manifest.sh` which tries to
find digests file using the plain arch names.
This results in a failure if you are building `aarch64` and `x86_64` together,
as it happens for example in `automation/release.sh`
Move the `format_archname` function to common.sh file, so that it can be also
used by `push-container-manifest.sh`

Signed-off-by: fossedihelm <ffossemo@redhat.com>