virt-controller: Keep iface requests until VM & VMI spec are patched
In the scenario virt-controller fails to patch the VMI spec with the requested
iface and VM is updated successfully, the request will be removed from the queue
(VM status.interfaceRequests) and the VMI will never be patched with the new
changes.

Remove interface requests when both the VM and VMI spec are patched following
interface requests.

Signed-off-by: Or Mergi <ormergi@redhat.com>