instancetype: Label ControllerRevisions with stored object metadata
ControllerRevisions are used to store point in time copies of instance
type and preference objects for each VirtualMachine making use of the
feature, see the following user-guide docs for more on this:

Instance types and preferences - Versioning
https://kubevirt.io/user-guide/virtual_machines/instancetypes/#versioning

This change adds labels to these ControllerRevision, exposing metadata
of the stored objects. With this in place it becomes easier to track the
objects stored within these ControllerRevisions and in the future
orchestrate the migration of these objects to newer versions of the API
and CRDs as these are introduced.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>