[e2e] Ignore LiveMigrate warning
EvictionStrategy can be set at cluster level.
When it is `LiveMigrate` and a vmi which is not live migratable is started,
a warning event is emitted. This warning can cause the failure of the test
because the utility function that wait for the vmi to start, generally,
does not tolerate any warning.
Set the specific warning in the warning ignore list when we detect that
the cluster level eviction strategy is `LiveMigrate`.

Signed-off-by: fossedihelm <ffossemo@redhat.com>