tests: leave some space for metadata on the backend PVC
In certain case, the targetcli backstores/fileio create command fails
with:
 Could not expand file to 1073741824 bytes

We can try to avoid this issue by creating a smaller backend image. We
simply hardcoded 800M instead of 1G as in these tests the size of the
disk doesn't matter. This is used to test the SCSI persistent
reservation ioctls.

Signed-off-by: Alice Frosi <afrosi@redhat.com>