tests, hotplug: Remove redundant migration
When testing multiple hotplug interfaces a redundant migration was
executed after the second hotplug.

By executing the migration before the new interface propagated to the
VMI, flakes are seen (pod interface missing on the target).

Therefore, drop one migration and depend on the one performed as part of
the `verifyDynamicInterfaceChange` helper.

Signed-off-by: Edward Haas <edwardh@redhat.com>