virt-controller, hotplug iface: Patch VMI after VM is updated
Make sure that interfaces in the VMI spec are set with MAC address
as the interfaces in the VM template spec.
So that the VMI will have the correct MAC addresses by the time
virt-launcher pod is created.

Without this change the guest interface may end up with random
MAC address instead of the once spesified in the VM spec.

Signed-off-by: Or Mergi <ormergi@redhat.com>