Quarantine flaky SEV test that tries to reset SEV capacity
The following test is very flaky[1][2]:

`AMD Secure Encrypted Virtualization (SEV) device management should
reset SEV capacity when the feature gate is disabled`

The timeout was previously increased[3] however that does not seem to
have resolved the flakiness.

Quarantining the test as it is causing impact to the sig-compute lanes.

[1] https://storage.googleapis.com/kubevirt-prow/reports/flakefinder/kubevirt/kubevirt/flakefinder-2023-06-27-168h.html#row1
[2] https://search.ci.kubevirt.io/?search=SEV+capacity+when+the+feature+gate+is+disabled&maxAge=336h&context=1&type=build-log&name=&excludeName=&maxMatches=5&maxBytes=20971520&groupBy=job
[3] https://github.com/kubevirt/kubevirt/pull/9907

Signed-off-by: Brian Carey <bcarey@redhat.com>