virt-controller, hotplug iface: Patch VMI interfaces MAC addresses
When iface is hotplugged to VM and KubeMacPool exist in the cluster,
KubeMacPool the VM spec interfaces and add a MAC address that is
monitored by KubeMacPool.

The requested interface MAC address wont replicate to the VMI spec,
causing the interface inside the guest to have a wrong MAC address.

To overcome this, virt-controller patch the VMI spec and add the MAC
addresses from the interfaces in the VM template spec.

Signed-off-by: Or Mergi <ormergi@redhat.com>