[e2e] Share KubeVirtDefaultConfig between all ginkgo nodes
InitRunConfiguration is executed in SynchronizedBeforeTestSetup function,
which is executed only in the process#1 (https://onsi.github.io/ginkgo/#parallel-suite-setup-and-cleanup-synchronizedbeforesuite-and-synchronizedaftersuite).
This causes that the TestRunConfiguration is misconfigured in the other processes, that
can result in failures in case of `LiveMigrate` cluster level eviction strategy.
Move the InitRunConfiguration function in `BeforeTestSuiteSetup` function.

Signed-off-by: fossedihelm <ffossemo@redhat.com>