Add new upgrade test script, using operator-sdk (#2410)
The CI current way to perform upgrade is deprecated. This PR add a new
script to perform the upgrade test in CI, using the operator SDK.

Assumptions:
1. The operator-sdk is available to use in the test
2. The bundle images of the last build of the current version, and the
   previous version are available at CI.
3. The output dir is configurable, to allow using the CI artifacts dir
   as the output dir (e.g. to store the SCC files).

Also, added the new `upgrade-test-operator-sdk` make target to run the
new script.
The script should run with the `INITIAL_VERSION` environment variable,
and optionally also with the the `OUTPUT_DIR` environment variable.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>