virt-handler, VMI iface status info source: Add pod info source
The new 'pod' info-source indicates whether an interface
is weird between the guest and the external interface (add by
the CNI), i.e.: in case of bridge binding interface, the existence
of the tap and bridge exist in virt-launcher pod.

The info source depends on checking in the network status volatile
cache, if the interface exist it means the network plumbing for it
finished and its ready to attach to the domain.

The indication whether an interface is weird to the guest in virt-launcher
(e.g.: tap and bridge exist) is the missing part that required for
virt-controller to clear interfaces with 'absent' state from the VMI spec.

Signed-off-by: Or Mergi <ormergi@redhat.com>