virt-handler, VMI iface status info-source: Add pod info source
The new 'pod' info-source indicates whether an interface
is wired between the guest and the external interface (add by
the CNI), i.e.: in case of bridge binding interface, the existence
of the tap and bridge exist in virt-launcher pod.

'pod' info source is set for interfaces that answers the following criterias:
- Not SR-IOV or Slirp.
- Exist in VMI spec (desired interfaces).
- Exist in Domain spec (attached to domain).
Meaning network setup phase 1 performed and these interface between the external
interface and the guest.

The indication whether an interface is wired to the guest in virt-launcher
(e.g.: tap and bridge exist) is the missing part that required for
virt-controller to clear interfaces with 'absent' state in the VMI spec.

Signed-off-by: Or Mergi <ormergi@redhat.com>