instancetype: Apply PreferredStorageClassName to StorageSpec if provided
DataVolumeSpec provides two mutually exclusive ways of defining
attributes of the destination volume through either a
PersistentVolumeClaimSpec [2] or StorageSpec [3].

Both provide ways of controlling the storageClass of the eventual PVC
but as outlined in bug #10025 previously PreferredStorageClassName was
only applied to the PersistentVolumeClaimSpec of a DataVolumeSpec.

This change corrects this ensuring PreferredStorageClassName is applied
to the StorageSpec if provided within a DataVolumeSpec.

[1] http://kubevirt.io/api-reference/main/definitions.html#_v1beta1_datavolumespec
[2] http://kubevirt.io/api-reference/main/definitions.html#_k8s_io_api_core_v1_persistentvolumeclaimspec
[3] http://kubevirt.io/api-reference/main/definitions.html#_v1beta1_storagespec

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>