pre-commit: Add config and use lint
This change introduces pre-commit, a useful framework that uses local
git hook scripts to ensure changes meet certain requirements *before*
they are fully committed.

The tool is opt-in and requires the user install pre-commit itself but
this is as easy as:

$ pip install --user pre-commit && pre-commit install

The provided config uses a number of standard hooks and additionally
runs the lint Makefile target.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>