Stop replacing "kubectl" with "kubectl virt" on virtctl commands
As of now, when executing virtctl create commands with the "--help"
argument, for example "virtctl create clone --help", it outputs
the following:

  # Create a manifest for a clone and use it to create a resource with kubectl virt:
  virtctl create clone --source-name sourceVM | kubectl virt create -f -

The above happens for other commands as well, e.g.
"virtctl create instancetype --help".

As of now, "kubectl" is being replaced with "kubectl virt", although
this is an invalid command that doesn't exist and cannot be used.

This commits stops replacing "kubectl" to "kubectl virt".

Signed-off-by: Itamar Holder <iholder@redhat.com>