virt-controller, vmi: Remove one abstraction layer for hot{un}plug
The handling of hotplug and hotunplug was centralized in a single method
which calculated the required patching and performed the path itself on
the pod object.

Removing this layer of abstraction allows the separation between the
calculation of the required patch and the actual patch.
This in turn also helps testing and the ability to patch only once the
pod (instead of the current state where the pod is patched multiple
times in a single reconcile loop).

Signed-off-by: Edward Haas <edwardh@redhat.com>