Remove tests_suite_test.go from focus-file set
If the file tests/tests_suite_test.go is included, it will pull in all
the other tests, which will void using -focus-file flag. Thus we extend
the grep to include suite.

We also add --dry-run flag in order to see which tests will actually
be run.

Finally we move the cluster-sync out of the inner loop, since we only
need to do this once at the start of the test cycle, since if a test
fails in a further cycle there's something wrong with the suite anyway.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>