virt-handler, SR-IOV hotplug: Filter attached interfaces
Interfaces who's external interface (added by CNI) exist in
virt-launcher pod but not attached to domain reported in the VMI status.

On SR-IOV interfaces hotplug flow, it's no longer enough to filter interfaces
according to existence in the VMI status.
It's necessary to also check they are not not attached to domain, otherwise
SR-IOV interfaces with only info-source 'multus-status' considered as attached.

Signed-off-by: Or Mergi <ormergi@redhat.com>