Stream guest console logs
Add `/spec/template/spec/domain/devices/logSerialConsole` API
to let the VM owner require the logs from the guest OS
over its serial console to be collected by the
the standard Kubernetes log architecture.

This will cause a `<log/>` element to be added to
the serial console device in the libvirt XML.

Logs will be collected to a private file (rotated by `virtlogd`).

Following the standard Kubernetes Streaming sidecar container logging pattern:
https://kubernetes.io/docs/concepts/cluster-administration/logging/#streaming-sidecar-container
a dedicated container will be eventaully used to stream the content of
the guest console log file to the Kubernetes logs so that they can be
collected/redirected to an external system as for the logs of a generic
containerized application.

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>