host-disk: Check existence of capacity and request storage resource
This is a follow up to #9810 where we started to limit the capacity of a
HostDisk to the request storage resource size if smaller than the
capacity storage resource size.

Both the requests and capacity are actually optional attributes of the
PersistentVolumeClaimInfo[1] object used to pass them into this code.
As such this change introduces some additional checks to ensure they are
present before attempting to populate the capacity of the HostDisk.

[1] https://github.com/kubevirt/kubevirt/blob/0f4cbbf0cf273c7014a16d5e09fb3d3dee77444a/staging/src/kubevirt.io/api/core/v1/types.go#L298-L304

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>