Support multiple watchdogs in the domain schema
Libvirt allows several watchdog devices since 9.1.0. The documentation
now states:

Having multiple watchdogs is usually not something very common, but be
aware that this might happen, for example, when an implicit watchdog
device is added as part of another device. For example the iTCO watchdog
being part of the ich9 southbridge, which is used with the q35 machine
type.

Signed-off-by: Vasiliy Ulyanov <vulyanov@suse.de>