virt-controller: Clear absent interfaces from spec
Currently, hot-unplugged interfaces are marked as absent and
remain in the VM & VMI spec even after the interface is no longer
exist in the guest or virt-launcher pod.

Remove detached interfaces with absent state from VM & VMI spec.

An interface is subject to be removed from sped if it exists in the
spec, have 'absent' state, and are not being reported in the VMI status.

In the scenario in-place hotplug is performed using Multus V4,
since there is no indication regarding the wiring and cache files Kubevirt
sets in virt-launcher pod (e.g.: bridge and tap device), the subject
interface may be removed without the cleaning up the wiring.
This will be addresses in follow up commit.

Signed-off-by: Or Mergi <ormergi@redhat.com>