Removing myself as code reviewer
My work assignments don't allow me to do repeated carefule code-review.
We have luckily now a solit set of reviewers which are aware of the main
architecture of kubevirt and can perform review.

I still plan to spend time on the approver role, to be able to perform
checks if the superficial palcements regarding architecture, API and so
forth are on the right place.

Signed-off-by: Roman Mohr <rmohr@google.com>