[e2e] Fix missing VMLiveUpdateFeaturesGate enabling
`VMLiveUpdateFeaturesGate` is enabled by default in `kubevirtresource.go`.
The e2e default feature gates are not set in every cluster, since if
`apply-default-e2e-configuration` is false, the default fg are not enabled,
resulting in a failure.
We should ensure, during the setup of the test, that the VMLiveUpdateFeaturesGate
is enabled.
Note that in the `EnableFeatureGate` function, there is a check that prevent the
patch if the fg is already present, thus this will not delay the execution time.

Signed-off-by: fossedihelm <ffossemo@redhat.com>