SRIOV hotplug, virt-controller: Don't update existing pod annotation
If an SRIOV interface is added to the spec of a running VM, don't update
the pod multus annotation with this interface.
To configure SRIOV on the pod, device plugin has to be invovled.
Kubelet invokes the device plugin of pod startup only.
Therefore, there is no point in adding the SRIOV interface to the multus
annotation at this point. It should be added only on migration to the
new target pod multus annotation.

Signed-off-by: Alona Paz <alkaplan@redhat.com>