SRIOV hotplug, virt-handler: Ignore SRIOV ifaces with no multus status
If there are SRIOV interfaces in the spec that are not in the domain
(plug pending)->
SRIOV hotplug is invoked.
This commit changes the code to invoke the SRIOV hotplug only if those
plug pending interfaces are reported in the multus annotation.

If they are not reported by the multus annotation it means the pod cannot
use the VF yet.

Signed-off-by: Alona Paz <alkaplan@redhat.com>