virt-controller: Introduce ShadowNode controller
This new controller watches shadowNode objects, extracts the kubevirt
controlled labels and annotations from it, then patches the k8s node
object.
The exception is the heartbeat "kubevirt.io/heartbeat" annotation that
will not be synced to the node by the controller.
The reason for this is to reduce patch frequency to the node. Keeping
this annotation inside the shadownode is enough for the purpose it
is currently serving.

Signed-off-by: Luboslav Pivarc <lpivarc@redhat.com>
Signed-off-by: Ram Lavi <ralavi@redhat.com>