Add Or() and IsNodeRunningVmis() predicates for TSC filtering
Add two new filter predicates that would be used to filter
nodes for TSC frequency calculations:

Or(): Can be used to express logical or between multiple
filters, e.g. Or(filter1, filter2).

IsNodeRunningVmis(vmiStore): returns a filter that includes
the node if VMIs are running on it.

This way nodes that are considered for TSC frequency calculations
need to be either schedulable or have VMIs running on it.

Signed-off-by: Itamar Holder <iholder@redhat.com>