hotplug interface: Fixing VM with default interface bug
If a VM has no interfaces, a default one may be automatically added to the
VMI.
A VM with no interfaces (that has a default interface added to its VMI)
ends up with only one interface after a hotplug although the VMI has two.
Therefore, after the VM is restarted, the default interface disappear.

This commit fixes the bug by explicitly adding the default interface to
the VM on hotplug.

Signed-off-by: Alona Paz <alkaplan@redhat.com>