[e2e] Fix migration bandwidth
In order to slow a migration, we set the `BandwidthPerMigration` to
a very low value.
Currently, since this value is a `Millibyte`, when converting it to byte,
it results that the migration params will be `BandwidthSet:false Bandwidth:0`,
which means that we are not slowing the migration.
Let's use `NewQuantity` instead of `NewMilliQuantity`, so that the result will be
`BandwidthSet:true Bandwidth:1`

Signed-off-by: fossedihelm <ffossemo@redhat.com>