[e2e] Relax prometheus cr operator test
Checking if the kubevirt PrometheusRule cr exists, a comparison
between the posted Cr and a generated one is done.
This comparison is too strict, since some vendors could use a
different `RUNBOOK_URL`, causing a failure.
From an operator functionality perspective, it is enough to check that the CR
exists and the group contains a set of rules.

Signed-off-by: fossedihelm <ffossemo@redhat.com>