[virtctl] Add support for an embedded novnc terminal
Embed a novnc instance in virtcl which can be started with

```
$ virtctl novnc myvm
```

In this commit, the connection url will only be printed:

```
Open http://127.0.0.1:38467?autoconnect=true to connect to the virtual machine.
```

By clicking on the provided link a browser window opens and the
presented NOVNC client connects right away to the VM.

The NOVNC instance is embedded as a zip file in virtctl and served
directly by reading from the zip archive with the helpf of the `zipfs`
package. As such no temporary space is needed and the binary size is
still kept relatively small since the client is embedded as a zip file.

This helps in two scenarios:
 * when users don't have a VNC client, or no compatible VNC client
   installed on the node.
 * when in addition users have to port-forward the connection. In that
   case directly hitting the port-forward address in the browser opens a
   terminal, instead of having to point a local VNC client to a TCP
   forward.

Signed-off-by: Roman Mohr <rmohr@google.com>