[virt-operator] cope with misscheduled virt-handler pods
Applying a custom `NoSchedule` taint to a nod will flip the `Available`
condition of KubeVirt to `False`:

```
  - lastProbeTime: "2023-08-07T14:15:43Z"
    lastTransitionTime: "2023-08-07T14:15:43Z"
    message: Deploying version devel with registry registry:5000/kubevirt
    reason: DeploymentInProgress
    status: "False"
    type: Available
```

This will only resolve back to `True` if the pod gets manually evicted,
or if the daemonset gets updated.

The background is, that we will see more up-to-date virt-handler pods in
ready state than we actually want.

Relax the readiness check slightly by counting misscheduled but
up-to-date and ready virt-handlers as something which does not trigger a
`DeploymentInProgress` state.

Signed-off-by: Roman Mohr <rmohr@google.com>