post-copy functest: tune stress and bandwidth limitation properly
Before this commit:
- The bandwidth limitation was 40Mi
- The stress size was 400M

A bandwidth of 40Mi is high. If in practice the network speed is
40Mi, the migration is expected to end within approximately
(400/40) == 10 seconds. However, it's entirely possible that
the actual network speed when CI is being run is much lower.

For example, if the actual bandwidth is 2Mi, it would take
200 seconds to complete. Since the migration is expected
to succeed after 180 seconds, the test would fail in this
scenario.

After this commit:
- The bandwidth limitation is 5Mi
- The stress size was 350M

This is a much more realistic tuning, since 1Mi is not a particularly
high network speed. This way, the migration is expected to finish
within approximately 70-80 seconds, and the timeout set in the test
is 150 seconds, which should be more than enough.

Signed-off-by: Itamar Holder <iholder@redhat.com>