runStressTest(): remove --timeout argument for stress-ng
Before this commit, running runStressTest(stressTimeoutSeconds=X)
would result in:
- Invoking stress-ng to run for X seconds.
- Test sleeping for X seconds to let the stress kick in.

While the second above makes sense, the first one doesn't.
What we usually want to test is migrating while stress
still runs in the background. Instead, before this commit
some memory is being dirtied, but stress stops as soon
as we start migrating which misses the whole point of
running a stress during a migration.

Signed-off-by: Itamar Holder <iholder@redhat.com>