Create a generic cache struct
The main idea is that when the `Get()` method is called, the cache
would either re-calculate the value or return the cached value
according to the parameters passed to it during its creation.

This cache gets a few parameters:
* `minRefreshDuration`: If the value is requested before the
  minRefreshDuration has passed, the cached value will be returned.
  If minRefreshDuration is zero, the value will always be recalculated.
* `useValueLock`: which hurts performance a bit, but makes the cache thread-safe.
* `reCalcFunc`: a function to re-calculate the desired value.

The value can be also explicitly set with `Set()`.

Signed-off-by: Itamar Holder <iholder@redhat.com>
(cherry picked from commit ccea92842c24b5718d0bf5d145ed5ca191e4bed0)