[virt-controller] Fix warning about ephemeral dir creation
The ephemeral directory for containerdisks needs to be known by
virt-controller to create the right pod templates. By mistake the
controller also tries to create the ephemeral directory.

By accident virt-controller also tries to create this directory in its
own image without actually using it. If that fails (e.g. permission
issues since untested and no functional difference) we get a warning
which can lead to confusions.

Removing the code which tries for no reason to create the directory
which look like this:

```
failed to create ephemeral disk dir: mkdir /var/run/kubevirt-ephemeral-disk
```

Signed-off-by: Roman Mohr <rmohr@google.com>