Reviewer Guide: Update guidelines around when what test is needed
I just tried to soften the language on what is required for a bug fix.
A bug fix does not require at least one functional test.
A bug fix requires an update to an existing e2e or unit test or a new e2e or unit test. It really depends on the bug.

Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>