Support getting dirty-rate as part of getDomainStats
This commit includes Dirty Rate calculation for 1
second with page-sampling mode.

Note that the page-sampling mode is the least accurate
on the one hand, but the most performant on the other.
It should not have any performance impact on the guest.
For more info: https://wiki.qemu.org/Features/DirtyRateCalc.

In addition, in order to get domain info with dirty rate
stats, the function will have to wait approximately 1 second
in order to let libvirt perform the calculation. This is
okay since domain stats are now cached, meaning they will
anyways not block the caller from receiving the domain stats.

Signed-off-by: Itamar Holder <iholder@redhat.com>