Allow image-upload to recover from PendingPopulation phase
This commit fixes virtctl image-upload command so we allow recovering from PendingPopulation phase, as it did with WaitForFirstConsumer.

This means that, if an image-upload fails due to a DV being in PendingPopulation phase, we can call the command again after creating a consumer and the upload will continue as expected.

Signed-off-by: Alvaro Romero <alromero@redhat.com>