Allow setting of requests/limits on VMExport pod
Currently the VMExport pod won't start inside a namespace with a
quota. It will also not work inside a namespace with a specific limit
range ratio. This commit allows one to specify the requests and limits
of the VMExport pod, and it adds some defaults so it will work inside
a namespace with a quota.

Signed-off-by: Alexander Wels <awels@redhat.com>