instancetype: Include object version in ControllerRevision name
This change is required by future changes that will introduce the
upgrade of stashed instance type and preference objects within
ControllerRevisions. As these objects are immutable within their
original ControllerRevison a new ControllerRevision must be created
containing the upgraded object.

As such we need to ensure the names of these resources are unique to
allow the original to only be deleted once the VirtualMachine has been
updated to point at the new ControllerRevision.

This change allows this by adding the version of the stashed object to
the name of newly generated ControllerRevisions.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>