instancetype: Add LatestVersion constant
Again, future changes will be introducing support for upgrading stashed
objects within ControllerRevisions.

We could work this out dynamically but that could easily cause issues
in the future if we were to ever look at introducing a v2{alpha,beta}
versions of the CRDs as we wouldn't want to upgrade to these until they
had fully GA'd.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>